requirements.txt
six
pandas
click
numpy
Pillow
tensorflow>=1.0.0